app/requirements.txt
streamlit
pandas
numpy
pdfplumber
python-docx
openpyxl
sentence-transformers
scikit-learn
faiss-cpu
nltk
